+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 #Teknoo Software - East Foundation - Change Log
 
+##[0.0.5-beta2] - 2017-08-
+###Updated
+- EastEndPointTrait use now StreamCallback to render a view and not call
+directly the view renderer
+
+###Fixed
+- Bad exception used in EastEndPointTrait for denied access exceptions, use now AccessDeniedHttpException.
+
 ##[0.0.5-beta1] - 2017-08-14
 ###Updated
 - Switch to PSR11, with PHP-DI as container